requirements.txt
black==23.12.1
chardet==5.2.0
click==8.1.7
dataclasses-json==0.6.3
Flask==3.0.0
GitPython==3.1.40
Jinja2==3.1.2
mypy==1.7.1
pathspec==0.11.2
pydantic==2.12.4
pygments==2.17.2
pylint==3.0.3
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
toml==0.10.2
types-PyYAML==6.0.12.20250915
